feat: mypage component 배치
